apis  more livewire recipes

+++ app/Livewire/Admin/Recipe/ShowRecipes.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Admin\Recipe;
 
 use App\Exports\RecipeExport;
 use App\Models\Recipe;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Livewire\WithPagination;
 use Livewire\Component;
@@ -15,8 +16,7 @@ class ShowRecipes extends Component
     public $search;
     public $openModal = false;
     public $name, $images,$description,$ingredients,$preparation,$identificador,$recipeId;
-    public $titleModal = "Crear Receta";
-    public $btnModal = "Crear";
+    public $titleModal = "Crear Receta", $btnModal = "Crear";
     public $rules = [
         'name'=> 'required',
         'images'=> 'required|image|mimes:png,jpg|max:2048',
@@ -26,6 +26,7 @@ class ShowRecipes extends Component
         // 'category' => 'required',
     ];
 
+    private $resetVariables = ['openModal','name','images','description','ingredients','preparation','btnModal','titleModal'];
 
     public function mount(){
         $this->identificador = rand(); //le asigna un numero al azar o random
@@ -38,17 +39,37 @@ class ShowRecipes extends Component
 
 
     public function abrirModal(){
-        $this->reset(['openModal','name','images','description','ingredients','preparation','btnModal','titleModal']);
+        $this->reset($this->resetVariables);
         $this->identificador = rand(); //le asigna un numero al azar o random (se hace para que input file cambie y no ponga el anterior)
         $this->openModal = true;
     }
 
     public function save(){
-        $this->validate();
         $images = $this->images->store('recipes');
-        Recipe::create(['name'=>$this->name,'images'=>$images, 'description'=>$this->description,'ingredients'=>$this->ingredients,'preparation'=>$this->preparation]);
-        $this->reset(['openModal','name','images','description','ingredients','preparation']);
-        $this->identificador = rand(); //le asigna un numero al azar o random (se hace para que input file cambie y no ponga el anterior)
+        $userAuth = Auth::user()->id;
+        $recipe = [
+            'name'=>$this->name,
+            'images'=>$images,
+            'description'=>$this->description,
+            'ingredients'=>$this->ingredients,
+            'preparation'=>$this->preparation,
+            'user_id'=> $userAuth
+        ];
+        if($this->btnModal=="Crear"){ 
+            $this->validate();
+            Recipe::create($recipe);
+            $this->reset($this->resetVariables);
+            $this->identificador = rand(); //le asigna un numero al azar o random (se hace para que input file cambie y no ponga el anterior)
+        }
+        elseif($this->btnModal=="Actualizar") { 
+            $recipeEdit = Recipe::find($this->recipeId); 
+            Storage::disk('public')->delete($recipeEdit->images); // Eliminar la imagen antigua
+            if($recipeEdit) {
+                $recipeEdit->update($recipe);
+                $this->reset($this->resetVariables);
+                $this->identificador = rand(); 
+            }
+        }
     }
 
     public function updated($propertyName){
@@ -56,10 +77,9 @@ class ShowRecipes extends Component
     }
     public function destroyRecipe(Recipe $recipe) {
         //eliminar imagen
-        if($recipe->images){
-            Storage::disk('public')->delete($recipe->images);
-        }
+        if($recipe->images){Storage::disk('public')->delete($recipe->images);}
         $recipe->delete();
+        $this->resetPage();
     }
     public function modalEdit(Recipe $recipe){
         $this->recipeId = $recipe->id;
